requirements.txt
gitdb==4.0.10
GitPython==3.1.37
smmap==5.0.1